requirements.txt
fastapi
groq
pydantic
python-dotenv
